feishu_bot/requirements.txt
lark-oapi>=1.4
PyYAML>=6.0
websockets>=12.0
requests>=2.28
